members display

+++ composeApp/src/commonMain/kotlin/com/lfssolutions/retialtouch/domain/model/dropdown/MemberType.kt
@@ -2,6 +2,6 @@ package com.lfssolutions.retialtouch.domain.model.dropdown
 
 data class MemberType(
     val id:Int=0,
-    val memberId:Int=0,
+    val memberId:Long=0,
     val name : String ="",
 )
