+++ composeApp/src/commonMain/kotlin/com/lfssolutions/retialtouch/domain/model/invoiceSaleTransactions/SaleInvoiceItem.kt
@@ -56,7 +56,7 @@ data class SaleInvoiceItem(
     @SerialName("locationId")
     val locationId: Long? = 0,
     @SerialName("memberId")
-    val memberId: Double? = null,
+    val memberId: Long? = null,
     @SerialName("memberName")
     val memberName: String? = null,
     @SerialName("paid")
